+++ src/OpenLoco/src/S5/S5.cpp
@@ -704,7 +704,7 @@ namespace OpenLoco::S5
             sub_4BAEC4();
             addr<0x0052334E, uint16_t>() = 0; // _thousandthTickCounter
             Gfx::invalidateScreen();
-            loadPreferredCurrencyAlways();
+            Scenario::loadPreferredCurrencyAlways();
             Gfx::loadCurrency();
             addr<0x00525F62, uint16_t>() = 0;
