+++ src/OpenLoco/src/Scenario.cpp
@@ -1,5 +1,6 @@
 #include "Scenario.h"
 #include "Audio/Audio.h"
+#include "Config.h"
 #include "Date.h"
 #include "Economy/Economy.h"
 #include "Entities/EntityManager.h"
@@ -21,6 +22,7 @@
 #include "MultiPlayer.h"
 #include "Objects/CargoObject.h"
 #include "Objects/ClimateObject.h"
+#include "Objects/ObjectIndex.h"
 #include "Objects/ObjectManager.h"
 #include "Objects/ScenarioTextObject.h"
 #include "OpenLoco.h"
@@ -478,5 +480,72 @@ namespace OpenLoco::Scenario
 
         args.push<uint16_t>(0);
         args.push<uint16_t>(0);
+    }
+
+    static loco_global<ObjectManager::SelectedObjectsFlags*, 0x50D144> _inUseobjectSelection;
+    static loco_global<ObjectManager::ObjectSelectionMeta, 0x0112C1C5> _objectSelectionMeta;
+
+    static std::span<ObjectManager::SelectedObjectsFlags> getInUseSelectedObjectFlags()
+    {
+        return std::span<ObjectManager::SelectedObjectsFlags>(*_inUseobjectSelection, ObjectManager::getNumInstalledObjects());
+    }
+
+    // 0x004C153B
+    void loadPreferredCurrencyAlways()
+    {
+        if (!Config::get().hasFlags(Config::Flags::preferredCurrencyAlways))
+        {
+            return;
+        }
+
+        const auto& preferredCurreny = Config::get().old.preferredCurrency;
+
+        if (preferredCurreny.isEmpty())
+        {
+            return;
+        }
+
+        ObjectManager::prepareSelectionList(false);
+        const auto oldCurrency = ObjectManager::getActiveObject(ObjectType::currency, getInUseSelectedObjectFlags());
+        if (oldCurrency.index != -1)
+        {
+            ObjectManager::selectObjectFromIndex(ObjectManager::SelectObjectModes::defaultDeselect, *oldCurrency.object._header, getInUseSelectedObjectFlags(), _objectSelectionMeta);
+        }
+        ObjectManager::selectObjectFromIndex(ObjectManager::SelectObjectModes::defaultSelect, preferredCurreny, getInUseSelectedObjectFlags(), _objectSelectionMeta);
+        ObjectManager::unloadUnselectedSelectionListObjects(getInUseSelectedObjectFlags());
+        ObjectManager::loadSelectionListObjects(getInUseSelectedObjectFlags());
+        ObjectManager::reloadAll();
+        Gfx::loadCurrency();
+        ObjectManager::freeSelectionList();
+    }
+
+    // 0x004C159C
+    void loadPreferredCurrencyNewGame()
+    {
+        if (!Config::get().hasFlags(Config::Flags::preferredCurrencyForNewGames))
+        {
+            loadPreferredCurrencyAlways();
+            return;
+        }
+
+        const auto& preferredCurreny = Config::get().old.preferredCurrency;
+
+        if (preferredCurreny.isEmpty())
+        {
+            return;
+        }
+
+        ObjectManager::prepareSelectionList(false);
+        const auto oldCurrency = ObjectManager::getActiveObject(ObjectType::currency, getInUseSelectedObjectFlags());
+        if (oldCurrency.index != -1)
+        {
+            ObjectManager::selectObjectFromIndex(ObjectManager::SelectObjectModes::defaultDeselect, *oldCurrency.object._header, getInUseSelectedObjectFlags(), _objectSelectionMeta);
+        }
+        ObjectManager::selectObjectFromIndex(ObjectManager::SelectObjectModes::defaultSelect, preferredCurreny, getInUseSelectedObjectFlags(), _objectSelectionMeta);
+        ObjectManager::unloadUnselectedSelectionListObjects(getInUseSelectedObjectFlags());
+        ObjectManager::loadSelectionListObjects(getInUseSelectedObjectFlags());
+        ObjectManager::reloadAll();
+        Gfx::loadCurrency();
+        ObjectManager::freeSelectionList();
     }
 }
