+++ src/OpenLoco/src/Scenario.h
@@ -147,4 +147,7 @@ namespace OpenLoco::Scenario
     void registerHooks();
     void formatChallengeArguments(const Objective& objective, const ObjectiveProgress& progress, FormatArguments& args);
     void sub_46115C();
+
+    void loadPreferredCurrencyAlways();
+    void loadPreferredCurrencyNewGame();
 }
